add sentry isolate
