+++ dart/lib/sentry.dart
@@ -35,3 +35,5 @@ export 'src/type_check_hint.dart';
 // exception extraction
 export 'src/exception_cause_extractor.dart';
 export 'src/exception_cause.dart';
+// Isolates
+export 'src/sentry_isolate.dart';
